remove script from function score bwc test and move to BasicBWCIT
Function score test was supposed to test that old nodes can still read
whatever a new node sends. However, currently there is no way to test
scripting outside the script packages.

Also, we don't really need to upgrade nodes for this test. We only need
a mixed cluster.

related to #13522
